+++ core/transaction/buy_coin.go
@@ -304,6 +304,9 @@ func (data BuyCoinData) Run(tx *Transaction, context *state.StateDB, isCheck boo
 		rewardPool.Add(rewardPool, tx.CommissionInBaseCoin())
 		context.AddBalance(sender, data.CoinToBuy, data.ValueToBuy)
 		context.SetNonce(sender, tx.Nonce)
+
+		context.DeleteCoinIfZeroReserve(data.CoinToBuy)
+		context.DeleteCoinIfZeroReserve(data.CoinToSell)
 	}
 
 	tags := common.KVPairs{