+++ core/transaction/sell_coin.go
@@ -313,6 +313,9 @@ func (data SellCoinData) Run(tx *Transaction, context *state.StateDB, isCheck bo
 		rewardPool.Add(rewardPool, tx.CommissionInBaseCoin())
 		context.AddBalance(sender, data.CoinToBuy, value)
 		context.SetNonce(sender, tx.Nonce)
+
+		context.DeleteCoinIfZeroReserve(data.CoinToBuy)
+		context.DeleteCoinIfZeroReserve(data.CoinToSell)
 	}
 
 	tags := common.KVPairs{